requirements.txt
-e .
aws-cdk.core
aws-cdk.aws-ecs
aws-cdk.aws-efs
aws-cdk.aws-ec2
aws-cdk.aws-logs
